requirements.txt
streamlit
altair
pandas
python-binance
scikit-learn
nest-asyncio
